requirements/requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
mysqlclient==2.1.1
PyMySQL==1.0.2
pandas==1.4.3
PyJWT==2.4.0
requests==2.28.1
